Added preference option to suppress handling of past events
It is not always desirable to pick up where you left off in terms of
fsevents. If you are watching a 'busy' folder you might bring down your
system just by trying to catch up with what happened while you were
gone. So while the application is able to handle events that
happened while it was not running this is now optional behaviour.